extract trip/place/untracked cards to /diary/cards
To tidy up /js/diary/, I am reorganizing it into subdirectories.
